Start the edge detector
Needs more tests (like, uh, actually detecting an edge).

With this edge detector set to a HOLD_COUNT of half the baud rate
(relative to the global clock rate) I think this can be used to align a
UART receiver when the falling edge of the start bit is detected.